Remove comparators from ChildrenEventsOption to ChildrenEventsOptionsManagerUtils.
