+++ src/main/java/ua/lviv/iot/childrenevents/manager/ChildrenEventsOptionsManagerUtils.java
@@ -10,21 +10,22 @@ import ua.lviv.iot.childrenevents.model.SortType;
 public class ChildrenEventsOptionsManagerUtils {
 
 	public static void sortByPrice(List<ChildrenEventsOption> childrenEventsOptions, SortType sortType) {
-		childrenEventsOptions.sort(new ChildrenEventsOption.SortByPriceInUAHComparator());
+		childrenEventsOptions.sort(new ChildrenEventsOptionsManagerUtils.SortByPriceInUAHComparator());
 		if (sortType == SortType.DESCENDING) {
 			Collections.reverse(childrenEventsOptions);
 		}
 	}
 
 	public static void sortByName(List<ChildrenEventsOption> childrenEventsOptions, SortType sortType) {
-		childrenEventsOptions.sort((new ChildrenEventsOption()).new SortByNameComparator());
+		childrenEventsOptions.sort((new ChildrenEventsOptionsManagerUtils()).new SortByNameComparator());
 		if (sortType == SortType.DESCENDING) {
 			Collections.reverse(childrenEventsOptions);
 		}
 
 	}
 
-	public static void sortByMaxQuantityOfChildren(List<ChildrenEventsOption> childrenEventsOptions,SortType sortType) {
+	public static void sortByMaxQuantityOfChildren(List<ChildrenEventsOption> childrenEventsOptions,
+			SortType sortType) {
 		childrenEventsOptions.sort(new Comparator<ChildrenEventsOption>() {
 			// @Override
 			public int compare(ChildrenEventsOption o1, ChildrenEventsOption o2) {
@@ -35,11 +36,29 @@ public class ChildrenEventsOptionsManagerUtils {
 			Collections.reverse(childrenEventsOptions);
 		}
 	}
-	
+
 	public static void sortByDurationInMinutes(List<ChildrenEventsOption> childrenEventsOptions, SortType sortType) {
-		childrenEventsOptions.sort((ChildrenEventsOption o1,ChildrenEventsOption o2)->o1.getDurationInMinutes()-o2.getDurationInMinutes());
+		childrenEventsOptions.sort((ChildrenEventsOption o1, ChildrenEventsOption o2) -> o1.getDurationInMinutes()
+				- o2.getDurationInMinutes());
 		if (sortType == SortType.DESCENDING) {
 			Collections.reverse(childrenEventsOptions);
+		}
+	}
+
+	public static class SortByPriceInUAHComparator implements Comparator<ChildrenEventsOption> {
+
+		@Override
+		public int compare(ChildrenEventsOption o1, ChildrenEventsOption o2) {
+			return (int) (o1.getPriceInUAH() - o2.getPriceInUAH());
+		}
+
+	}
+
+	public class SortByNameComparator implements Comparator<ChildrenEventsOption> {
+
+		@Override
+		public int compare(ChildrenEventsOption ceo1, ChildrenEventsOption ceo2) {
+			return (int) (ceo1.getName().compareTo(ceo2.getName()));
 		}
 	}
 }
