+++ src/main/java/ua/lviv/iot/childrenevents/model/ChildrenEventsOption.java
@@ -1,7 +1,5 @@
 package ua.lviv.iot.childrenevents.model;
 
-import java.util.Comparator;
-
 public class ChildrenEventsOption {
 	protected String name;
 	protected String contacts;
@@ -71,7 +69,6 @@ public class ChildrenEventsOption {
 	public ChildrenEventsOption(double priceInUAH) {
 		this(null, null, priceInUAH, 0, 0, null);
 	}
-	
 
 	public ChildrenEventsOption() {
 		this(0);
@@ -128,30 +125,5 @@ public class ChildrenEventsOption {
 				+ ", maxQuantityOfChildren=" + maxQuantityOfChildren + ", durationInMinutes=" + durationInMinutes
 				+ ", venue=" + venue + "]";
 	}
-///SortByPriceInUAHComparator
-	public static class SortByPriceInUAHComparator implements Comparator<ChildrenEventsOption> {
 
-		@Override
-		public int compare(ChildrenEventsOption o1, ChildrenEventsOption o2) {
-			return (int) (o1.getPriceInUAH() - o2.getPriceInUAH());
-		}
-
-	}
-
-	public class SortByNameComparator implements Comparator<ChildrenEventsOption> {
-
-		@Override
-		public int compare(ChildrenEventsOption ceo1, ChildrenEventsOption ceo2) {
-			return (int) (ceo1.getName().compareTo(ceo2.getName()));
-		}
-	}
-	
-	/*public static Comparator<ChildrenEventsOption> priceInUAHComparator = new Comparator<ChildrenEventsOption>() {
-		//@Override
-		public int compare(ChildrenEventsOption ceo1, ChildrenEventsOption ceo2) {
-			return (ceo2.getPriceInUAH() < ceo1.getPriceInUAH()) ? -1
-					: ceo2.getPriceInUAH() == ceo1.getPriceInUAH() ? 0 : 1;
-		}
-	};*/
-	
 }
