+++ src/test/java/ua/lviv/iot/childrenevents/manager/BaseChildrenEventsOptionsManagerTest.java
@@ -13,7 +13,6 @@ public abstract class BaseChildrenEventsOptionsManagerTest {
 
 	public void createChildrenEventsOptions() {
 		childrenEventsOptions = new LinkedList<ChildrenEventsOption>();
-	//	childrenEventsOptions.add(new ChildrenEventsOption(name, contacts, priceInUAH, maxQuantityOfChildren, durationInMinutes, venue)
 		childrenEventsOptions.add(new ChildrenEventsOption("aaa", null, 10, 7, 120, EventVenue.MIXED));
 		childrenEventsOptions.add(new ChildrenEventsOption("bab", null, 11, 17, 240, EventVenue.MIXED));
 		childrenEventsOptions.add(new ChildrenEventsOption("baa", null, 150, 8, 90, EventVenue.INDOOR));
